backend/requirements.txt
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
distlib==0.3.8
Django==5.1.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
dnspython==2.7.0
dominate==2.9.1
email_validator==2.2.0
fastapi==0.115.9
fastapi-cli==0.0.7
filelock==3.16.1
Flask==3.1.1
Flask-Bootstrap==3.3.7.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
flask-marshmallow==1.2.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
Flask-WTF==1.2.1
greenlet==3.2.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpickle==3.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.21.3
marshmallow-sqlalchemy==1.0.0
mdurl==0.1.2
mysql-connector==2.2.9
mysql-connector-python==9.0.0
mysql-connector-python-rf==2.2.2
mysqlclient==2.2.4
orjson==3.10.15
packaging==24.1
pillow==10.4.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==1.10.13
pydantic-extra-types==2.10.2
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.41
sqlparse==0.5.1
starlette==0.45.3
typer==0.15.2
typing_extensions==4.14.0
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.34.0
virtualenv==20.26.6
visitor==0.1.3
waitress==3.0.1
watchfiles==1.0.4
websockets==15.0
Werkzeug==3.1.3
whitenoise==6.8.2
WTForms==3.1.2
